requirements.txt
numpy==1.26.4
opencv-python-headless==4.11.0.86
PyYAML==6.0.2
tflite-runtime==2.14.0
paho-mqtt
requests
onnxruntime